Move themeSelection into own file
